# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "SprintPark API",
      "description": "API documentation for SprintPark website",
      "version": "1.0.0"
    },
    "host": "127.0.0.1:5202",
    "schemes": ["http"],
    "paths": {
      "/signup": {
        "post": {
          "summary": "User Signup",
          "description": "Registers a new user with the provided details.",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "description": "User signup details",
              "schema": {
                "type": "object",
                "required": ["username", "email", "password", "full_name", "designation", "reporting_manager", "employee_id", "mobile_number", "location", "date_of_birth", "blood_group"],
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" },
                  "full_name": { "type": "string" },
                  "designation": { "type": "string" },
                  "reporting_manager": { "type": "string" },
                  "employee_id": { "type": "string" },
                  "mobile_number": { "type": "string" },
                  "location": { "type": "string" },
                  "date_of_birth": { "type": "string", "format": "date" },
                  "blood_group": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "201": { "description": "User signed up successfully." },
            "400": { "description": "Invalid request or missing fields." },
            "409": { "description": "User already exists." }
          }
        }
      },
     "/signin": {
  "post": {
    "summary": "User Signin",
    "description": "Allows users to sign in with their username and password.",
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "parameters": [
      {
        "in": "body",
        "name": "body",
        "required": true,
        "description": "User credentials",
        "schema": {
          "type": "object",
          "required": ["username", "password"],
          "properties": {
            "username": { "type": "string" },
            "password": { "type": "string", "format": "password" }
          }
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Signin successful.",
        "schema": {
          "type": "object",
          "properties": {
            "message": { "type": "string", "example": "Signin successful." },
            "token": { "type": "string", "example": "jwt_token" },
            "full_name": { "type": "string", "example": "User Full Name" },
            "designation": { "type": "string", "example": "User Designation" },
            "reporting_manager": { "type": "string", "example": "Reporting Manager" },
            "employee_id": { "type": "string", "example": "Employee ID" },
            "email": { "type": "string", "format": "email", "example": "user@example.com" }
          }
        }
      },
      "400": { "description": "Invalid JSON format or missing fields." },
      "404": { "description": "User does not exist." },
      "403": { "description": "Account is not active." },
      "401": { "description": "Invalid password." }
    }
  }
},
      "/getAllEmployees": {
        "get": {
          "summary": "Get All Employees",
          "description": "Retrieves a list of all employees.",
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "A list of employees.",
              "schema": {
                "type": "object",
                "properties": {
                  "employees": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "username": { "type": "string" },
                        "email": { "type": "string", "format": "email" },
                        "full_name": { "type": "string" },
                        "designation": { "type": "string" },
                        "reporting_manager": { "type": "string" },
                        "employee_id": { "type": "string" },
                        "mobile_number": { "type": "string" },
                        "location": { "type": "string" },
                        "date_of_birth": { "type": "string", "format": "date" },
                        "blood_group": { "type": "string" },
                        "status": { "type": "string" }
                      }
                    }
                  }
                }
              }
            },
            "500": { "description": "Database connection failed." }
          }
        }
      }
    }
  }
  